chore: mock 데이터 삭제, 불필요한 주석 삭제
