+++ src/domains/course/pages/CourseLearnClientPage.tsx
@@ -38,7 +38,6 @@ export default function CourseLearnClient({ enrollmentId }: CourseLearnClientPro
 
   return (
     <div className={styles['course-learn']}>
-      {/* 헤더 */}
       <header className={styles['course-learn__header']}>
         <div className={styles['course-learn__header-inner']}>
           <div className={styles['course-learn__header-left']}>
@@ -58,7 +57,6 @@ export default function CourseLearnClient({ enrollmentId }: CourseLearnClientPro
       </header>
 
       <div className={styles['course-learn__layout']}>
-        {/* 메인 플레이어 */}
         <main className={styles['course-learn__main']}>
           <div className={styles['course-learn__player-wrap']}>
             {currentLecture.type === 'VIDEO' &&
@@ -97,7 +95,6 @@ export default function CourseLearnClient({ enrollmentId }: CourseLearnClientPro
             )}
           </div>
 
-          {/* 현재 강의 정보 */}
           <div className={styles['course-learn__info-card']}>
             <div className={styles['course-learn__info-row']}>
               {currentLecture.type === 'VIDEO' ? (
@@ -117,7 +114,6 @@ export default function CourseLearnClient({ enrollmentId }: CourseLearnClientPro
           </div>
         </main>
 
-        {/* 사이드바 - 커리큘럼 */}
         <aside className={styles['course-learn__aside']}>
           <div className={styles['course-learn__curriculum']}>
             <h3 className={styles['course-learn__curriculum__title']}>커리큘럼</h3>
